requirements.txt
streamlit
requests
python-dotenv
streamlit-audiorec
SpeechRecognition
pydub
